+++ packages/components/organisms/f-header/CHANGELOG.md
@@ -2,6 +2,22 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v10.9.0
+------------------------------
+*November 1, 2022*
+
+### Fixed
+- JE Logo vertical positioning.
+- Country text & logo alignment in country selector
+
+### Changed
+- Friendly Name Truncation Tweak Points
+
+### Added
+- `showCorporateLink` - Shows "Corporate Ordering" Link.
+- `isCondensed` - A flag to hide icons, reduce spacing between nav-links on mid size, reduce container padding on tweak points and apply stricter friendly name truncation.
+
+
 v10.8.0
 ------------------------------
 *November 1, 2022*